Add final repositories and edit-service
