+++ pages/api/controllers/transactions/edit.js
@@ -1,4 +1,4 @@
-const db = require("../../db");
+import { editTransactionService } from "../../services/edit-service";
 
 export default async (req, res) => {
 	// TODO
@@ -14,7 +14,7 @@ export default async (req, res) => {
 	try {
 		const { originalRow, editedRow } = req.body;
 
-		const result = await db.editTransaction(originalRow, editedRow);
+		const result = await editTransactionService(originalRow, editedRow);
 
 		if (result === "OK") {
 			console.log("ok!");
