+++ pages/api/services/edit-service.js
@@ -1,0 +1,6 @@
+import { editTransaction } from "../repositories/edit-repository";
+
+export const editTransactionService = async (originalRow, editedRow) => {
+	const results = await editTransaction(originalRow, editedRow);
+	return results;
+};
